# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: DuFense GRC Platform API Gateway
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /privacy/dsars:
    get:
      operationId: listDsars
      tags: [privacy]
      summary: List Data Subject Access Requests
      responses:
        "200":
          description: DSARs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dsar"
    post:
      operationId: createDsar
      tags: [privacy]
      summary: Create a DSAR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DsarInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dsar"
  /privacy/dsars/{id}:
    patch:
      operationId: updateDsar
      tags: [privacy]
      summary: Update a DSAR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DsarUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dsar"
  /privacy/dpias:
    get:
      operationId: listDpias
      tags: [privacy]
      summary: List DPIAs
      responses:
        "200":
          description: DPIAs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dpia"
    post:
      operationId: createDpia
      tags: [privacy]
      summary: Create a DPIA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DpiaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dpia"
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List all service desk tickets
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
components:
  schemas:
...
    Dsar:
      type: object
      required: [id, dsarId, type, subject, received, due, status]
      properties:
        id:
          type: integer
        dsarId:
          type: string
        type:
          type: string
        subject:
          type: string
        received:
          type: string
        due:
          type: string
        status:
          type: string
        daysLeft:
          type: ["integer", "null"]
    DsarInput:
      type: object
      required: [type, subject, due]
      properties:
        type:
          type: string
        subject:
          type: string
        due:
          type: string
    DsarUpdate:
      type: object
      properties:
        status:
          type: string
        daysLeft:
          type: integer
    Dpia:
      type: object
      required: [id, dpiaId, name, risk, status, owner, updated]
      properties:
        id:
          type: integer
        dpiaId:
          type: string
        name:
          type: string
        risk:
          type: string
        status:
          type: string
        owner:
          type: string
        updated:
          type: string
    DpiaInput:
      type: object
      required: [name, risk, owner]
      properties:
        name:
          type: string
        risk:
          type: string
        owner:
          type: string
    Ticket:
      type: object
      required: [id, ticketId, priority, title, category, assignee, status, sla, createdAt]
      properties:
        id:
          type: integer
        ticketId:
          type: string
        priority:
          type: string
        title:
          type: string
        category:
          type: string
        assignee:
          type: string
        status:
          type: string
        sla:
          type: string
        createdAt:
          type: string
    TicketInput:
      type: object
      required: [priority, title, category, assignee, sla]
      properties:
        priority:
          type: string
        title:
          type: string
        category:
          type: string
        assignee:
          type: string
        sla:
          type: string
    TicketUpdate:
      type: object
      properties:
        priority:
          type: string
        title:
          type: string
        status:
          type: string
        assignee:
          type: string